Move code to parse the PersonIdent structure from a tag or commit object into PersonIdent module.
